% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{load_pack}
\alias{load_pack}
\title{Load a parameter pack}
\usage{
load_pack(id)
}
\arguments{
\item{id}{Pack identifier (as reported by \code{\link[=list_packs]{list_packs()}}), or a path to a
YAML file.}
}
\value{
An object of class \code{gi_pack}.
}
\description{
Load a parameter pack
}
\examples{
pack <- load_pack("ercp_acute_cholangitis")
pack$title
}
\seealso{
\code{\link[=list_packs]{list_packs()}}, \code{\link[=validate_pack]{validate_pack()}}, \code{\link[=scenario]{scenario()}}
}
