% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{pack_search_path}
\alias{pack_search_path}
\title{Directories searched for parameter packs}
\usage{
pack_search_path()
}
\value{
Character vector of existing directories, in search order.
}
\description{
Directories searched for parameter packs
}
\examples{
pack_search_path()
}
\seealso{
\link{parameter-packs}
}
